pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "flask-migrate>=4.1.0",
    "python-bidi>=0.6.6",
    "arabic-reshaper>=3.0.0",
    "openpyxl>=3.1.5",
    "requests>=2.32.5",
    "python-dateutil>=2.9.0.post0",
    "pytest>=8.4.1",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.14.1",
    "factory-boy>=3.3.3",
    "coverage>=7.10.5",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "sendgrid>=6.12.4",
    "flask-wtf>=1.2.2",
    "apscheduler>=3.11.0",
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.29.0",
]
